@@ -2,9 +2,9 @@
   "name": "react-webpack",
   "version": "1.0.0",
   "description": "generate webpack react demo",
-  "main": "index.js",
+  "main": "js/index.js",
   "scripts": {
-    "start": "webpack-dev-server --hot --progress --colors",
+    "start": "webpack-dev-server --hot --progress --profile --colors",
     "build": "webpack --progress --colors"
   },
   "repository": {
@@ -18,17 +18,35 @@
   "author": "yixuan",
   "license": "MIT",
   "devDependencies": {
-    "babel-core": "^5.8.25",
-    "babel-loader": "^5.3.2",
-    "css-loader": "^0.12.1",
-    "react": "^0.13.3",
-    "react-hot-loader": "^1.3.0",
-    "react-router": "^0.13.3",
+    "babel-core": "^6.18.2",
+    "babel-loader": "^6.2.8",
+    "babel-preset-es2015": "^6.13.2",
+    "babel-preset-react": "^6.11.1",
+    "babel-preset-stage-0": "^6.5.0",
+    "css-loader": "^0.26.0",
+    "react": "^15.4.1",
+    "react-hot-loader": "^3.0.0-beta.6",
+    "react-router": "^3.0.0",
     "webpack": "^1.12.2",
     "webpack-dev-server": "^1.11.0"
   },
   "dependencies": {
-    "lodash": "~3.10.1",
+    "lodash": "~4.17.2",
+    "react-bootstrap": "^0.30.7",
+    "react-dom": "^15.4.1",
     "react-kendo": "~0.13.11"
+  },
+  "babel": {
+    "presets": [
+      "es2015",
+      "stage-0",
+      "react"
+    ],
+    "plugins": [
+      "react-hot-loader/babel"
+    ]
+  },
+  "jest": {
+    "rootDir": "./js"
   }
 }
