@@ -1,25 +1,49 @@
-var webpack = require('webpack');
+var webpack = require('webpack')
+
+const HOST = process.env.HOST || 'localhost'
+const PORT = process.env.PORT || '8080'
+
+var loaders = [
+  {
+    test: /\.jsx?$/,
+    exclude: /node_modules/,
+    loaders: ['babel']
+  }
+]
+
 module.exports = {
-    entry: [
-      'webpack/hot/only-dev-server',
-      "./js/app.js"
-    ],
-    output: {
-        path: './build',
-        filename: "bundle.js"
-    },
-    module: {
-        loaders: [
-            { test: /\.js?$/, loaders: ['react-hot', 'babel'], exclude: /node_modules/ },
-            { test: /\.js$/, exclude: /node_modules/, loader: 'babel-loader'},
-            { test: /\.css$/, loader: "style!css" },
-            {test: /\.less/,loader: 'style-loader!css-loader!less-loader'}
-        ]
-    },
-    resolve:{
-        extensions:['','.js','.json']
-    },
-    plugins: [
-      new webpack.NoErrorsPlugin()
-    ]
-};
+  entry: [
+    'react-hot-loader/patch',
+    'webpack/hot/only-dev-server',
+    './js/index.js'
+  ],
+  devtool: process.env.WEBPACK_DEVTOOL || 'cheap-module-source-map',
+  output: {
+    path: 'dist',
+    publicPath: 'dist',
+    filename: 'bundle.js'
+  },
+  resolve: {
+    extensions: ['', '.js', '.jsx']
+  },
+  module: {
+    loaders
+  },
+  devServer: {
+    contentBase: '.',
+    noInfo: true,
+    hot: true,
+    inline: true,
+    historyApiFallback: true,
+    host: HOST,
+    port: PORT
+  },
+  plugins: [
+    new webpack.NoErrorsPlugin(),
+    new webpack.HotModuleReplacementPlugin(),
+    new webpack.ProvidePlugin({
+      $: 'jquery',
+      jQuery: 'jquery'
+    })
+  ]
+}
